pyproject.toml
[tool.poetry]
name        = "money-transfer-starter"
version     = "0.1.0"
description = "Enterprise money-transfer micro-services stack"
authors     = ["Your Name <you@example.com>"]
readme      = "README.md"
packages    = [{ include = "common" }, { include = "api_gateway" },
               { include = "user_service" }, { include = "transaction_service" }]

[tool.poetry.dependencies]
python = "^3.12"

# ------------- Web / API -------------
fastapi                         = "^0.111"
uvicorn                         = { extras = ["standard"], version = "^0.29" }
httpx                           = "^0.27"
strawberry-graphql              = { extras = ["fastapi"], version = "^0.233" }

# ------------- gRPC / Protobuf -------------
grpcio                          = "^1.64"

# ------------- Async infra -------------
redis                           = "^5.0"
aiokafka                        = "^0.10"
sqlalchemy                      = "^2.0"
asyncpg                         = "^0.29"
casbin                          = "^1.30"
slowapi                         = "^0.1.8"
tenacity                        = "^8.3"

# ------------- Auth / JWT -------------
PyJWT                           = "^2.8"
cachetools                      = "^5.3"

# ------------- Observability -------------
prometheus-fastapi-instrumentator = "^6.1"
opentelemetry-sdk               = "^1.27"
opentelemetry-exporter-otlp     = "^1.27"

# ------------- Misc -------------
pydantic                        = "^2.7"

[tool.poetry.group.dev.dependencies]
pytest       = "^8.2"
grpcio-tools                    = "^1.64"
black        = "^24.2"
isort        = "^5.13"
mypy         = "^1.10"

[build-system]
requires      = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
